refactor: CoreDataManager 구조체를 CoreDataManageable 프로토콜로 변경, 그에 따른 수정
